Refactor MediaInfo and create MediaInfoUtil helper class.
